refactor: Refactor to separate view and logic of comment component
